*优化包接收器。 *解决碎包处理问题 *其他若干修改
git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@168 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/network/message_handler.hpp
@@ -38,7 +38,7 @@ public:
 	std::string name;
 	MessageID msgID;
 	MessageArgs* pArgs;
-	int16 msgLen;					// 如果长度为-1则为非固定长度消息
+	int32 msgLen;					// 如果长度为-1则为非固定长度消息
 	
 	virtual void handle(MemoryStream& s)
 	{
